pyproject.toml
[tool.poetry]
name = "bustail_core"
version = "0.1.0"
description = ""
authors = ["alex <alexanderdemure@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.58.1"
uvicorn = "^0.11.5"
fastapi-sqlalchemy = "^0.1.4"
psycopg2 = "^2.8.5"
alembic = "^1.4.2"
databases = "^0.3.2"
asyncpg = "^0.20.1"

python-jose = "^3.1.0"
passlib = "^1.7.2"
bcrypt = "^3.1.7"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
